requirements.txt
mtcnn==1.0.0
pillow==11.1.0
numpy==2.2.4
tensorflow==2.19.0
keras==3.9.0
keras-facenet==0.3.2
scipy==1.15.2
opencv-python==4.11.0.86
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
mlxtend==0.23.4
tensorflow==2.19.0
deepface==0.0.93
tf-keras==2.19.0
python-imagesearch==1.3.0